separated web and mobile apps
